+++ CHANGELOG.md
@@ -5,7 +5,11 @@
 ## 2023-04-18: v0.1 alpha
 
 First steps setting the bot. Testing *commandhandling* and different types of sent objects. The **bot** is in
-*bot.py* file but use a set of classes. 
+*bot.py* file but use a set of classes. List of commands:  
+
+- **/start**: to say hi.
+- **/language**: to select language.
+- **/help**: to ask for help about how to use the bot.
 
 Reach **musiCal bot** [here](https://t.me/caltools_bot).
 Feel free to contact me by [mail](mailto:rodrigovalla@protonmail.ch) or reach me in